feat: introduce a new common runtime
