+++ packages/typescript-axios-runtime/package.json
@@ -29,6 +29,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@nahkies/typescript-common-runtime": "workspace:^",
     "qs": "^6.14.0",
     "tslib": "^2.8.1"
   },
