+++ packages/typescript-common-runtime/README.md
@@ -1,0 +1,10 @@
+# @nahkies/typescript-common-runtime
+
+[![CI/CD](https://github.com/mnahkies/openapi-code-generator/actions/workflows/ci.yml/badge.svg)](https://github.com/mnahkies/openapi-code-generator/actions?query=branch%3Amain+event%3Apush)
+[![npm](https://img.shields.io/npm/dm/%40nahkies%2Ftypescript-common-runtime.svg)](https://www.npmjs.com/package/@nahkies/typescript-common-runtime)
+
+This is a supporting package for code generated using [@nahkies/openapi-code-generator](https://www.npmjs.com/package/@nahkies/openapi-code-generator) shared across the typescript templates.
+
+You can [read the docs](https://openapi-code-generator.nahkies.co.nz/) to find out more!
+
+It's not intended by be used standalone. Similar in spirit to [tslib](https://www.npmjs.com/package/tslib)
