+++ packages/typescript-common-runtime/jest.config.js
@@ -1,0 +1,12 @@
+const base = require("../../jest.base")
+const {name: displayName} = require("./package.json")
+
+/**
+ * @type { import('@jest/types').Config.ProjectConfig }
+ */
+const config = {
+  ...base,
+  displayName,
+}
+
+module.exports = config
