+++ packages/typescript-common-runtime/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@nahkies/typescript-koa-runtime",
+  "name": "@nahkies/typescript-common-runtime",
   "version": "0.22.0",
-  "description": "Runtime package for code generated by @nahkies/openapi-code-generator using the typescript-koa template",
+  "description": "Runtime package for code generated by @nahkies/openapi-code-generator using the typescript templates",
   "license": "MIT",
   "author": {
     "name": "Michael Nahkies",
@@ -11,7 +11,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/mnahkies/openapi-code-generator.git",
-    "directory": "packages/typescript-koa-runtime"
+    "directory": "packages/typescript-common-runtime"
   },
   "bugs": {
     "url": "https://github.com/mnahkies/openapi-code-generator/issues"
@@ -22,25 +22,20 @@
       "import": "./dist/errors.js",
       "types": "./dist/errors.d.ts"
     },
-    "./server": {
-      "require": "./dist/server.js",
-      "import": "./dist/server.js",
-      "types": "./dist/server.d.ts"
+    "./validation": {
+      "require": "./dist/validation.js",
+      "import": "./dist/validation.js",
+      "types": "./dist/validation.d.ts"
     },
-    "./joi": {
-      "require": "./dist/joi.js",
-      "import": "./dist/joi.js",
-      "types": "./dist/joi.d.ts"
+    "./types": {
+      "require": "./dist/types.js",
+      "import": "./dist/types.js",
+      "types": "./dist/types.d.ts"
     },
-    "./zod-v3": {
-      "require": "./dist/zod-v3.js",
-      "import": "./dist/zod-v3.js",
-      "types": "./dist/zod-v3.d.ts"
-    },
-    "./zod-v4": {
-      "require": "./dist/zod-v4.js",
-      "import": "./dist/zod-v4.js",
-      "types": "./dist/zod-v4.d.ts"
+    "./request-bodies/url-search-params": {
+      "require": "./dist/request-bodies/url-search-params.js",
+      "import": "./dist/request-bodies/url-search-params.js",
+      "types": "./dist/request-bodies/url-search-params.d.ts"
     }
   },
   "scripts": {
@@ -49,25 +44,14 @@
     "test": "jest"
   },
   "dependencies": {
-    "@nahkies/typescript-common-runtime": "workspace:^",
     "tslib": "^2.8.1"
   },
   "peerDependencies": {
-    "@koa/cors": "^5.0.0",
-    "@koa/router": "^13.0.0 || ^14.0.0",
     "joi": "^17.13.3 || ^18.0.1",
-    "koa": "^2.16.1 || ^3.0.0",
-    "koa-body": "^6.0.1",
     "zod": "^3.25.74 || ^4.1.12"
   },
   "peerDependenciesMeta": {
-    "@koa/cors": {
-      "optional": true
-    },
     "joi": {
-      "optional": true
-    },
-    "koa-body": {
       "optional": true
     },
     "zod": {
@@ -75,15 +59,9 @@
     }
   },
   "devDependencies": {
-    "@koa/cors": "^5.0.0",
-    "@koa/router": "^14.0.0",
-    "@types/koa": "^3.0.1",
-    "@types/koa__cors": "^5.0.1",
-    "@types/koa__router": "^12.0.5",
+    "@jest/globals": "^30.2.0",
     "jest": "^30.2.0",
     "joi": "^18.0.2",
-    "koa": "^3.1.1",
-    "koa-body": "^6.0.1",
     "typescript": "^5.9.3",
     "zod": "^3.25.74"
   },
@@ -96,10 +74,9 @@
   ],
   "keywords": [
     "@nahkies/openapi-code-generator",
+    "openapi",
     "runtime",
-    "typescript-koa",
-    "koa",
-    "zod"
+    "typescript"
   ],
   "publishConfig": {
     "access": "public"
