+++ packages/typescript-express-runtime/package.json
@@ -32,11 +32,6 @@
       "import": "./dist/joi.js",
       "types": "./dist/joi.d.ts"
     },
-    "./zod": {
-      "require": "./dist/zod.js",
-      "import": "./dist/zod.js",
-      "types": "./dist/zod.d.ts"
-    },
     "./zod-v3": {
       "require": "./dist/zod-v3.js",
       "import": "./dist/zod-v3.js",
@@ -55,6 +50,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@nahkies/typescript-common-runtime": "workspace:^",
     "tslib": "^2.8.1"
   },
   "peerDependencies": {
