--- packages/typescript-express-runtime/src/zod.ts
@@ -1,2 +1,0 @@
-/** @deprecated Import zod-v3 / zod-v4 instead */
-export * from "./zod-v3"
