+++ packages/typescript-fetch-runtime/package.json
@@ -49,6 +49,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@nahkies/typescript-common-runtime": "workspace:^",
     "qs": "^6.14.0",
     "tslib": "^2.8.1"
   },
